Move all goals to goals package
Just makes it so the ai package doesn't fill up with custom goals
